made the start up a little more modular see notes
OK....  So I kind of layed the ground work for several things.
I created some globals that are the base line options for how it runs.  at the top of frist tab you can set fullscreen or not and move or not. you dont have to comment or uncomment the different objects. I also seperated out the setup update CV code and the visuals.  I also created a frame object that just has info on the current size of the frame to use in the gbcv class.