new class for choosing video input
Now only lines 55 and 57 need to be (un)commented to switch between the video inputs.

Your videoStartupManager fuction is not contained in the inputVideo class as the loadFrames method.

Iswebcam is handled automatically within the constructors, of which there are 2 - one for the Movie input version and one for the Capture version.

I can work on synching to what you have done last night as this version is not quite up to date with your visuals seperation work.

For now, just run this example and see what you think about it.